refactor simulation function creation
