+++ pyabc/inference/__init__.py
@@ -1,0 +1,1 @@
+from .smc import ABCSMC
